Move ParameterizedSamePointConstraint to geometry/optimization/iris_internal.h.
