+++ geometry/optimization/iris_internal.h
@@ -60,6 +60,46 @@ class SamePointConstraint : public solvers::Constraint {
       nullptr};
 };
 
+class ParameterizedSamePointConstraint : public solvers::Constraint {
+  DRAKE_NO_COPY_NO_MOVE_NO_ASSIGN(ParameterizedSamePointConstraint);
+
+  ParameterizedSamePointConstraint(
+      const multibody::MultibodyPlant<double>* plant,
+      const systems::Context<double>& context,
+      const std::function<Eigen::VectorXd(const Eigen::VectorXd&)>&
+          parameterization_double,
+      const std::function<AutoDiffVecXd(const AutoDiffVecXd&)>&
+          parameterization_autodiff,
+      int parameterization_dimension);
+
+  ~ParameterizedSamePointConstraint() override;
+
+  void set_frameA(const multibody::Frame<double>* frame) {
+    same_point_constraint_.set_frameA(frame);
+  }
+
+  void set_frameB(const multibody::Frame<double>* frame) {
+    same_point_constraint_.set_frameA(frame);
+  }
+
+ private:
+  void DoEval(const Eigen::Ref<const Eigen::VectorXd>& x,
+              Eigen::VectorXd* y) const override;
+
+  void DoEval(const Eigen::Ref<const AutoDiffVecXd>& x,
+              AutoDiffVecXd* y) const override;
+
+  void DoEval(const Eigen::Ref<const VectorX<symbolic::Variable>>& x,
+              VectorX<symbolic::Expression>* y) const override;
+
+  geometry::optimization::internal::SamePointConstraint same_point_constraint_;
+  const std::function<Eigen::VectorXd(const Eigen::VectorXd&)>&
+      parameterization_double_;
+  const std::function<AutoDiffVecXd(const AutoDiffVecXd&)>&
+      parameterization_autodiff_;
+  int parameterization_dimension_;
+};
+
 /* Defines a MathematicalProgram to solve the problem
  min_q (q-d) CᵀC (q-d)
  s.t. setA in frameA and setB in frameB are in collision in q.
